+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - GH-116 More friendly error objects (Domenic Denicola)
 - GH-115 Client hangs on server "hard kills" (i.e., RST)
+- GH-111 JSON parser only works on objects (not arrays)
 - GH-110 emit expectContinue (Paul Bouzakis)
 - Fix "undefined" log message in string_client.js
 - GH-107